add salt and features
Add salt and features.
Need to work on PasswordReset and need to add links to the new html pages in the old once such as change_password.html, password_reset.html, set_new_pwd.html.
Need to change the css code on the new files to look like the old once.